requirements.txt
python-telegram-bot==22.5
pymongo==4.15.3
openai==1.59.5
requests==2.32.5
beautifulsoup4==4.14.2
python-dotenv==1.1.1
selenium
webdriver-manager
aiohttp
fastapi==0.115.0
uvicorn[standard]==0.32.0
